create requirement:
1. assign roll_no to student form
   - roll_no :program_year-major_code-sequence by "total _marks desc "
2. after promotion , assign exam 'total_avg_score' to 'total_marks' in education.student.final.result
3. update roll_no to student form by education.student.final.result

+++ models/education_exam/exam_results.py
@@ -118,7 +118,7 @@ class ExamOverallResultsLine(models.Model):
     student_id = fields.Many2one('education.student',string = 'Student')
     state = fields.Selection([('draft','Draft'),('done','Completed')], string='State',required=True, default='draft',track_visibility='onchange')
     result_ids = fields.One2many('education.subject.overallresults','overall_result_id',string='Result Ids')
-    total_avg_score = fields.Float(string='Total Average Score', readonly= True)
+    total_avg_score = fields.Integer(string='Total Average Score', readonly= True)
     pass_or_fail = fields.Boolean(string='Pass/Fail')
    
     def get_name(self):
